refactor : Not Found Exception
