+++ projecthub/src/main/java/com/project/hub/exceptions/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
-package com.project.hub.exception;
+package com.project.hub.exceptions;
 
-import com.project.hub.exception.exception.BusinessException;
+import com.project.hub.exceptions.exception.BusinessException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
