+++ projecthub/src/main/java/com/project/hub/exceptions/exception/NotFoundException.java
@@ -1,0 +1,10 @@
+package com.project.hub.exceptions.exception;
+
+import com.project.hub.exceptions.ExceptionCode;
+
+public class NotFoundException extends BusinessException{
+
+  public NotFoundException(ExceptionCode exceptionCode) {
+    super(exceptionCode);
+  }
+}
